corrigiendo la clase compraFrutas y algunas correcciones menores

+++ JS/funciones.js
@@ -77,7 +77,7 @@ function compraFruta(codigo, cantidad){
         if(fruta.codigo === instancia.codigo){
             band1 = 1
             if(confirm(`Su pedido ya incluye ${fruta.cantidadKg} Kg de ${recuperaFruta(fruta.codigo).nombre} ¿Desea agregar otros ${instancia.cantidadKg} Kg ?`)){
-                fruta.cantidadKg = parseInt(fruta.cantidadKg) + parseInt(instancia.cantidadKg)
+                fruta.cantidadKg = parseFloat(fruta.cantidadKg) + parseFloat(instancia.cantidadKg)
             }else{
                 console.warn("Su pedido no fue modificado.")
             }
